requirements.txt
scrapy>=2.11.0
scrapy-playwright>=0.0.34
playwright>=1.40.0
itemadapter>=0.8.0
supabase>=2.0.0
python-dotenv>=1.0.0
